Combine commands that can be combined, move all commands to Commands namespace
