+++ Commands/BanCmds.cs
@@ -1,16 +1,162 @@
-﻿using static Cliptok.Helpers.BanHelpers;
+using static Cliptok.Helpers.BanHelpers;
 
 namespace Cliptok.Commands
 {
-    class Bans
+    public class BanCmds
     {
+        [Command("ban")]
+        [Description("Bans a user from the server, either permanently or temporarily.")]
+        [AllowedProcessors(typeof(SlashCommandProcessor))]
+        [RequireHomeserverPerm(ServerPermLevel.Moderator), RequirePermissions(DiscordPermission.BanMembers)]
+        public async Task BanSlashCommand(SlashCommandContext ctx,
+            [Parameter("user"), Description("The user to ban")] DiscordUser user,
+            [Parameter("reason"), Description("The reason the user is being banned")] string reason,
+            [Parameter("keep_messages"), Description("Whether to keep the users messages when banning")] bool keepMessages = false,
+            [Parameter("time"), Description("The length of time the user is banned for")] string time = null,
+            [Parameter("appeal_link"), Description("Whether to show the user an appeal URL in the DM")] bool appealable = false,
+            [Parameter("compromised_account"), Description("Whether to include special instructions for compromised accounts")] bool compromisedAccount = false
+        )
+        {
+            // Initial response to avoid the 3 second timeout, will edit later.
+            var eout = new DiscordInteractionResponseBuilder().AsEphemeral(true);
+            await ctx.DeferResponseAsync(true);
+
+            // Edits need a webhook rather than interaction..?
+            DiscordWebhookBuilder webhookOut = new();
+            int messageDeleteDays = 7;
+            if (keepMessages)
+                messageDeleteDays = 0;
+
+            if (user.IsBot)
+            {
+                webhookOut.Content = $"{Program.cfgjson.Emoji.Error} To prevent accidents, I won't ban bots. If you really need to do this, do it manually in Discord.";
+                await ctx.EditResponseAsync(webhookOut);
+                return;
+            }
+
+            DiscordMember targetMember;
+
+            try
+            {
+                targetMember = await ctx.Guild.GetMemberAsync(user.Id);
+                if ((await GetPermLevelAsync(ctx.Member)) == ServerPermLevel.TrialModerator && ((await GetPermLevelAsync(targetMember)) >= ServerPermLevel.TrialModerator))
+                {
+                    webhookOut.Content = $"{Program.cfgjson.Emoji.Error} As a Trial Moderator you cannot perform moderation actions on other staff members.";
+                    await ctx.EditResponseAsync(webhookOut);
+                    return;
+                }
+            }
+            catch
+            {
+                // do nothing :/
+            }
+
+            TimeSpan banDuration;
+            if (time is null)
+                banDuration = default;
+            else
+            {
+                try
+                {
+                    banDuration = HumanDateParser.HumanDateParser.Parse(time).Subtract(ctx.Interaction.CreationTimestamp.DateTime);
+                }
+                catch
+                {
+                    webhookOut.Content = $"{Program.cfgjson.Emoji.Error} There was an error parsing your supplied ban length!";
+                    await ctx.EditResponseAsync(webhookOut);
+                    return;
+                }
+
+            }
+
+            DiscordMember member;
+            try
+            {
+                member = await ctx.Guild.GetMemberAsync(user.Id);
+            }
+            catch
+            {
+                member = null;
+            }
+
+            if (member is null)
+            {
+                await BanHelpers.BanFromServerAsync(user.Id, reason, ctx.User.Id, ctx.Guild, messageDeleteDays, ctx.Channel, banDuration, appealable);
+            }
+            else
+            {
+                if (DiscordHelpers.AllowedToMod(ctx.Member, member))
+                {
+                    if (DiscordHelpers.AllowedToMod(await ctx.Guild.GetMemberAsync(ctx.Client.CurrentUser.Id), member))
+                    {
+                        await BanHelpers.BanFromServerAsync(user.Id, reason, ctx.User.Id, ctx.Guild, messageDeleteDays, ctx.Channel, banDuration, appealable);
+                    }
+                    else
+                    {
+                        webhookOut.Content = $"{Program.cfgjson.Emoji.Error} I don't have permission to ban **{DiscordHelpers.UniqueUsername(user)}**!";
+                        await ctx.EditResponseAsync(webhookOut);
+                        return;
+                    }
+                }
+                else
+                {
+                    webhookOut.Content = $"{Program.cfgjson.Emoji.Error} You don't have permission to ban **{DiscordHelpers.UniqueUsername(user)}**!";
+                    await ctx.EditResponseAsync(webhookOut);
+                    return;
+                }
+            }
+            reason = reason.Replace("`", "\\`").Replace("*", "\\*");
+            if (banDuration == default)
+                await ctx.Channel.SendMessageAsync($"{Program.cfgjson.Emoji.Banned} {user.Mention} has been banned: **{reason}**");
+            else
+                await ctx.Channel.SendMessageAsync($"{Program.cfgjson.Emoji.Banned} {user.Mention} has been banned for **{TimeHelpers.TimeToPrettyFormat(banDuration, false)}**: **{reason}**");
+
+            webhookOut.Content = $"{Program.cfgjson.Emoji.Success} User was successfully bonked.";
+            await ctx.EditResponseAsync(webhookOut);
+        }
+
+        [Command("unban")]
+        [Description("Unbans a user who has been previously banned.")]
+        [AllowedProcessors(typeof(SlashCommandProcessor), typeof(TextCommandProcessor))]
+        [HomeServer, RequireHomeserverPerm(ServerPermLevel.Moderator), RequirePermissions(permissions: DiscordPermission.BanMembers)]
+        public async Task UnbanCmd(CommandContext ctx, [Description("The user to unban, usually a mention or ID")] DiscordUser targetUser, [Description("Used in audit log only currently")] string reason = "No reason specified.")
+        {
+            if ((await Program.db.HashExistsAsync("bans", targetUser.Id)))
+            {
+                await UnbanUserAsync(ctx.Guild, targetUser, $"[Unban by {DiscordHelpers.UniqueUsername(ctx.User)}]: {reason}");
+                await ctx.RespondAsync($"{Program.cfgjson.Emoji.Unbanned} Successfully unbanned **{DiscordHelpers.UniqueUsername(targetUser)}**.");
+            }
+            else
+            {
+                bool banSuccess = await UnbanUserAsync(ctx.Guild, targetUser);
+                if (banSuccess)
+                    await ctx.RespondAsync($"{Program.cfgjson.Emoji.Unbanned} Successfully unbanned **{DiscordHelpers.UniqueUsername(targetUser)}**.");
+                else
+                {
+                    await ctx.RespondAsync($"{Program.cfgjson.Emoji.Error} {ctx.Member.Mention}, that user doesn't appear to be banned, *and* an error occurred while attempting to unban them anyway.\nPlease contact the bot owner if this wasn't expected, the error has been logged.");
+                }
+            }
+        }
+
+        [Command("baninfo")]
+        [Description("Show information about the ban for a user.")]
+        [AllowedProcessors(typeof(SlashCommandProcessor))]
+        [RequireHomeserverPerm(ServerPermLevel.TrialModerator)]
+        [RequirePermissions(DiscordPermission.ModerateMembers)]
+        public async Task BanInfoSlashCommand(
+            SlashCommandContext ctx,
+            [Parameter("user"), Description("The user whose ban information to show.")] DiscordUser targetUser,
+            [Parameter("public"), Description("Whether to show the output publicly. Default: false")] bool isPublic = false)
+        {
+            await ctx.RespondAsync(embed: await BanHelpers.BanStatusEmbed(targetUser, ctx.Guild), ephemeral: !isPublic);
+        }
+
         [Command("massbantextcmd")]
         [TextAlias("massban", "bigbonk")]
         [AllowedProcessors(typeof(TextCommandProcessor))]
         [HomeServer, RequireHomeserverPerm(ServerPermLevel.Moderator)]
         public async Task MassBanCmd(TextCommandContext ctx, [RemainingText] string input)
         {
-
             List<string> usersString = input.Replace("\n", " ").Replace("\r", "").Split(' ').ToList();
             List<ulong> users = usersString.Select(x => Convert.ToUInt64(x)).ToList();
             if (users.Count == 1 || users.Count == 0)
@@ -217,30 +363,5 @@ namespace Cliptok.Commands
             else
                 await ctx.Channel.SendMessageAsync($"{Program.cfgjson.Emoji.Banned} {targetMember.Mention} has been banned for **{TimeHelpers.TimeToPrettyFormat(banDuration, false)}**: **{reason}**");
         }
-
-        [Command("unbantextcmd")]
-        [TextAlias("unban")]
-        [Description("Unbans a user who has been previously banned.")]
-        [AllowedProcessors(typeof(TextCommandProcessor))]
-        [HomeServer, RequireHomeserverPerm(ServerPermLevel.Moderator), RequirePermissions(permissions: DiscordPermission.BanMembers)]
-        public async Task UnbanCmd(TextCommandContext ctx, [Description("The user to unban, usually a mention or ID")] DiscordUser targetUser, [Description("Used in audit log only currently")] string reason = "No reason specified.")
-        {
-            if ((await Program.db.HashExistsAsync("bans", targetUser.Id)))
-            {
-                await UnbanUserAsync(ctx.Guild, targetUser, $"[Unban by {DiscordHelpers.UniqueUsername(ctx.User)}]: {reason}");
-                await ctx.RespondAsync($"{Program.cfgjson.Emoji.Unbanned} Successfully unbanned **{DiscordHelpers.UniqueUsername(targetUser)}**.");
-            }
-            else
-            {
-                bool banSuccess = await UnbanUserAsync(ctx.Guild, targetUser);
-                if (banSuccess)
-                    await ctx.RespondAsync($"{Program.cfgjson.Emoji.Unbanned} Successfully unbanned **{DiscordHelpers.UniqueUsername(targetUser)}**.");
-                else
-                {
-                    await ctx.RespondAsync($"{Program.cfgjson.Emoji.Error} {ctx.Member.Mention}, that user doesn't appear to be banned, *and* an error occurred while attempting to unban them anyway.\nPlease contact the bot owner if this wasn't expected, the error has been logged.");
-                }
-            }
-        }
-
     }
 }
