+++ Commands/ReminderCmds.cs
@@ -1,6 +1,6 @@
-﻿namespace Cliptok.Commands
+namespace Cliptok.Commands
 {
-    public class Reminders
+    public class ReminderCmds
     {
         public class Reminder
         {
@@ -71,6 +71,5 @@
             await Program.db.ListRightPushAsync("reminders", JsonConvert.SerializeObject(reminderObject));
             await ctx.RespondAsync($"{Program.cfgjson.Emoji.Success} I'll try my best to remind you about that on <t:{TimeHelpers.ToUnixTimestamp(t)}:f> (<t:{TimeHelpers.ToUnixTimestamp(t)}:R>)"); // (In roughly **{TimeHelpers.TimeToPrettyFormat(t.Subtract(ctx.Message.Timestamp.DateTime), false)}**)");
         }
-
     }
 }
