+++ Commands/SlowmodeCmds.cs
@@ -1,6 +1,6 @@
-﻿namespace Cliptok.Commands.InteractionCommands
+namespace Cliptok.Commands
 {
-    internal class SlowmodeInteractions
+    public class SlowmodeCmds
     {
         [Command("slowmode")]
         [Description("Slow down the channel...")]
@@ -73,6 +73,5 @@
                 }
             }
         }
-
     }
 }
