+++ Commands/UserNoteCmds.cs
@@ -2,8 +2,17 @@ using static Cliptok.Helpers.UserNoteHelpers;
 
 namespace Cliptok.Commands.InteractionCommands
 {
-    internal class UserNoteInteractions
+    internal class UserNoteCmds
     {
+        [Command("Show Notes")]
+        [SlashCommandTypes(DiscordApplicationCommandType.UserContextMenu)]
+        [AllowedProcessors(typeof(UserCommandProcessor))]
+        [RequireHomeserverPerm(ServerPermLevel.TrialModerator), RequirePermissions(DiscordPermission.ModerateMembers)]
+        public async Task ShowNotes(UserCommandContext ctx, DiscordUser targetUser)
+        {
+            await ctx.RespondAsync(embed: await UserNoteHelpers.GenerateUserNotesEmbedAsync(targetUser), ephemeral: true);
+        }
+
         [Command("note")]
         [Description("Manage user notes")]
         [AllowedProcessors(typeof(SlashCommandProcessor))]
